Move logic.app_commands to command
