+++ src/commands/config.py
@@ -3,7 +3,7 @@ import discord
 from embed import ErrorEmbed
 from embeds.config.permissions import ConfigPermissionsView
 from embeds.config.sources import ConfigSourcesView
-from logic.app_commands import SimpleCommand, SimpleCommandGroup
+from command import SimpleCommand, SimpleCommandGroup
 from logic.config import user_is_admin, user_is_admin_or_has_config_permissions
