+++ src/commands/distribution.py
@@ -1,7 +1,7 @@
 from discord import Interaction, app_commands
 
 from embeds.distribution import DistributionEmbed
-from logic.app_commands import SimpleCommand
+from command import SimpleCommand
 from logic.distribution import distribution
 from logic.color import UserColor
 from logic.roll import Advantage
