+++ src/commands/initiative.py
@@ -2,7 +2,7 @@ import discord
 
 from embeds.initiative import InitiativeContainerView
 from logic.initiative import InitiativeTracker
-from logic.app_commands import SimpleCommand
+from command import SimpleCommand
 from logic.voice_chat import VC, SoundType
