+++ src/context_menus/timestamp.py
@@ -1,7 +1,7 @@
 import discord
 from embed import SimpleEmbed
 from embeds.timestamp import RelativeTimestampEmbed
-from logic.app_commands import SimpleContextMenu
+from command import SimpleContextMenu
 from logic.timestamp import get_relative_timestamp_from_message
