+++ src/logic/roll.py
@@ -2,7 +2,7 @@ import dataclasses
 from enum import Enum
 import d20
 
-from logic.app_commands import ChoicedEnum
+from command import ChoicedEnum
 
 
 class Advantage(ChoicedEnum):
